adding weekends work
