+++ Dockerfile
@@ -1,86 +1,51 @@
-FROM python:latest
+FROM continuumio/miniconda3
 MAINTAINER vsochat@stanford.edu
+
+# docker build -t vanessa/tunel .
 
 RUN apt-get update 
 RUN apt-get -y install build-essential
-RUN apt-get -y install apt-utils
-RUN apt-get -y install cmake
-RUN apt-get -y install wget
-RUN apt-get -y install unzip
-RUN apt-get -y install python-tk
-RUN apt-get -y install python-dev 
-RUN apt-get -y install pkg-config
-RUN apt-get -y install libffi-dev 
-RUN apt-get -y install libssl-dev
-RUN apt-get -y install qt-sdk
-RUN apt-get -y install libtool \
-                       autotools-dev \
-                       automake \
-                       autoconf \
-                       git
+RUN apt-get -y install apt-utils cmake wget unzip libffi-dev libssl-dev \
+                       libtool autotools-dev automake autoconf git \
+                       libarchive-dev squashfs-tools uuid-dev \
+                       vim jq aria2 nginx
 
-
-# Software dependencies. 
-# This is equivalent of requirements.txt, run on image build
-RUN pip install --upgrade pip
-RUN pip install cycler==0.10.0
-RUN pip install h5py==2.6.0
-RUN pip install matplotlib==1.5.3
-RUN pip install nose==1.3.7
-RUN pip install numpy==1.11.0
-RUN pip install pandas==0.18.1
-RUN pip install Pillow==3.2.0
-RUN pip install pydicom==0.9.9
-RUN pip install pyparsing==2.1.10
-RUN pip install python-dateutil==2.5.3
-RUN pip install pytz==2016.4
-RUN pip install runcython==0.2.5
-RUN pip install scipy==0.18.1
-RUN pip install six==1.10.0
-RUN pip install simplejson
-
-# Dependencies for simple web app demo
-RUN pip install Flask==0.10.1
-RUN pip install Flask-SQLAlchemy==2.0
-RUN pip install Jinja2==2.7.3
-RUN pip install MarkupSafe==0.23
-RUN pip install SQLAlchemy==0.9.9
-RUN pip install Werkzeug==0.10.4
-RUN pip install gunicorn==19.3.0
-RUN pip install itsdangerous==0.24
-RUN pip install flask-restful
-RUN pip install singularity
-RUN pip install butterfly
-
-RUN butterfly.server.py --generate-certs --host="localhost" 
-RUN butterfly.server.py --generate-user-pkcs=$USER
-
-# Install Singularity
-
-WORKDIR /tmp
-RUN git clone http://www.github.com/singularityware/singularity
-WORKDIR /tmp/singularity
-RUN git checkout -b development
-RUN git pull origin development
-RUN ./autogen.sh
-RUN ./configure --prefix=/usr/local 
-RUN make 
-RUN make install
-
-
-# Make directories for code and data
+ENV DEBIAN_FRONTEND noninteractive
+ENV PATH /opt/conda/bin:$PATH
 RUN mkdir /code
 RUN mkdir /data
 
-# Add the code
+# Sregistry with pull fix
+RUN git clone -b update/pull https://www.github.com/singularityhub/sregistry-cli && \
+                 cd sregistry-cli && python setup.py install
+
+WORKDIR /tmp
+RUN wget https://github.com/singularityware/singularity/releases/download/2.4.3/singularity-2.4.3.tar.gz \
+    && tar -xzf singularity-2.4.3.tar.gz && cd singularity-2.4.3 && ./autogen.sh && ./configure --prefix=/usr/local \
+    && make && make install
+
 ADD . /code
-RUN ./code/data/bases/generate_bases.sh
+
+# Set up nginx
+RUN cp /code/script/nginx.conf /etc/nginx/nginx.conf && \
+    cp /code/script/nginx.gunicorn.conf /etc/nginx/sites-enabled/default && \
+    chmod u+x /code/script/entrypoint.sh && \
+    cp /code/tunel/config_dummy.py /code/tunel/config.py && \
+    chmod u+x /code/script/generate_key.sh && \
+    /bin/bash /code/script/generate_key.sh /code/tunel/config.py
+
+RUN /opt/conda/bin/pip install --upgrade pip && \
+    /opt/conda/bin/pip install -r /code/requirements.txt
+
+# Install HPC Container Maker
+#RUN git clone https://github.com/NVIDIA/hpc-container-maker.git && cd hpc-container-maker && python setup.py install
 
 # Clean up
-RUN apt-get autoremove -y
-RUN apt-get clean
-RUN rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/*
+RUN apt-get autoremove -y && \
+    apt-get clean && \
+    rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/*
 
-EXPOSE 8000
-
+ENTRYPOINT ["/bin/bash", "/code/script/entrypoint.sh"]
 WORKDIR /code
+EXPOSE 5000
+EXPOSE 80
