+++ data/README.md
@@ -1,3 +1,4 @@
 # Data
 
-The data folder is for operating system bases (bases) and user built containers (containers). The bases folder is intended to be used to issue commands to retrieve information about the packages available. If the bases aren't generated, the user is prompted to do so.
+This data folder will be mapped to the Singularity Registry database in the container to ensure
+that containers are saved if the application is started and stopped.
