--- nginx.conf
@@ -1,17 +1,0 @@
-server {
-
-    listen 80;
-    server_name example.org;
-    charset utf-8;
-
-    location /static {
-        alias /code/app/static;
-    }
-
-    location / {
-        proxy_pass http://web:5000;
-        proxy_set_header Host $host;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    }
-}
