+++ requirements.txt
@@ -1,12 +1,17 @@
-flask
+spython==0.0.22
+ipython
+paramiko==2.4.0
+cycler==0.10.0
+pyparsing==2.1.10
+Flask==0.10.1
+flask-cors
+Flask-SQLAlchemy==2.0
+Flask-WTF
+Jinja2==2.7.3
+MarkupSafe==0.23
+SQLAlchemy==0.9.9
+Werkzeug==0.10.4
+gunicorn==19.3.0
+itsdangerous==0.24
 flask-restful
-pandas
-requests
-retrying
-selenium
-simplejson
-scikit-learn
-pygments
-google-api-python-client
-google
-oauth2client==3.0
+shelljob
