+++ script/README.md
@@ -1,0 +1,5 @@
+The [docker-compose.yml](docker-compose.yml) was first used for the application,
+but deprecated in favor of a single container. The author (@vsoch) believes
+that it is more likely for a user to be comfortable running the container directly
+without needing to clone the Github repository and installing Docker compose. 
+And so this was a choice taken for usability of this portal.
