+++ script/nginx.gunicorn.conf
@@ -1,0 +1,21 @@
+server {
+    listen 80;
+
+    server_name _;
+
+    location /static {
+        alias /code/tunel/static;
+    }
+
+    location / {
+        proxy_pass         http://127.0.0.1:5000/;
+        proxy_redirect     off;
+
+        proxy_set_header   Host                 $host;
+        proxy_set_header   X-Real-IP            $remote_addr;
+        proxy_connect_timeout 900s;
+        proxy_read_timeout 900s;
+        proxy_set_header   X-Forwarded-For      $proxy_add_x_forwarded_for;
+        proxy_set_header   X-Forwarded-Proto    $scheme;
+    }
+}
