+++ tunel/static/css/style.css
@@ -21,12 +21,9 @@ form {
   background: #fff;
   padding: 4em 4em 2em;
   max-width: 420px;
-  margin: 20px auto 0;
-  box-shadow: 0 0 1em #999 !important;
   border-radius: 2px;
 }
 .question_button {
-  box-shadow: 0 0 1em #999 !important;
   border-radius:2px; 
   width:100% !important; 
   margin-top:5px
@@ -62,12 +59,10 @@ label {
 
 form input[type="submit"] {
   background: #e32929;
-  border-radius: 2px;
   border: none;
   color: #fff;
   cursor: pointer;
   display: block;
-  font-size: 1em;
   line-height: 1.6em;
   margin: 2em 0 0;
   outline: none;
