+++ tunel/templates/base.html
@@ -14,8 +14,19 @@
       <script src="{{ url_for('static', filename='js/jquery.min.js') }}"></script>
       <script src="{{ url_for('static', filename='js/select2.min.js') }}"></script>
       <link rel="stylesheet" href="{{ url_for('static', filename='css/select2.min.css') }}">
-      <script src="{{ url_for('static', filename='js/bootstrap.min.js') }}"></script>
       <script src="{{ url_for('static', filename='js/datatables.min.js') }}"></script>
+
+      <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" 
+              integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" 
+              crossorigin="anonymous"></script>
+      <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" 
+              integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
+              crossorigin="anonymous"></script>
+      <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" 
+              integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" 
+              crossorigin="anonymous"></script>
+      <link rel="stylesheet" 
+            href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"  crossorigin="anonymous">
 
       <link href="https://fonts.googleapis.com/css?family=Open+Sans+Condensed:300" rel="stylesheet">
       <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
@@ -28,10 +39,9 @@
   </head>
   <body>
 
-    <div class="container-fluid">
+    {% include "messages.html" %}
     {% block content %}
     {% endblock %}
-    </div>
 
 {% block scripts %}
 {% endblock %}
