+++ tunel/templates/index.html
@@ -5,28 +5,59 @@
 {% endblock %}
     
 {% block content %}
-
-    <div class="col-md-3">
-
-        <div class="row">
-         <a href="/browser" target="_blank"><button class="btn btn-default btn-xl question_button">File Browser</button></a>
+<div class="container-fluid">
+    <div class="row">
+        <div class="col-md-2" style='padding-top:50px'>
+        {% include "sidebar.html" %}
         </div>
+        <div class="col-md-10" style="padding-top:50px">
 
-        <div class="row">
-         <a href="/containers" target="_blank"><button class="btn btn-default btn-xl question_button" disabled>Containers</button></a>
+        <table id="label_table" class="table table-striped table-bordered" cellspacing="0" width="100%">
+        <thead>
+            <tr>
+                <th>Location</th>
+                <th>Client</th>
+                <th>Container</th>
+            </tr>
+        </thead>
+        <tfoot>
+            <tr>
+                <th>Location</th>
+                <th>Client</th>
+                <th>Container</th>
+            </tr>
+        </tfoot>
+        <tbody>
+            {% for image in images %}
+            <tr>
+                <td>{{ image.location() }}</td>
+                <td>{{ image.client }}</td>
+                <td>{{ image.uri }}</td>
+            </tr>
+            {% endfor %}
+        </tbody>
+        </table>
+        {% if images %}{% else %}
+           <p class="alert alert-info">Hey friend! You don't have any containers yet. Why don't you 
+              <a href="/pull">pull</a> some?</p>
+        {% endif %}
+
         </div>
-
-        <div class="row">
-         <a href="/recipe" target="_blank"><button class="btn btn-default btn-xl question_button" disabled>Recipe Generator</button></a>
-        </div>
-
-</div>
-
-<div class ="col-md-9">
-<img style="width:350px;padding-top:180px;margin-left:100px" src="{{ url_for('static', filename='img/robot.png') }}">
-<h1 style="color:#999;padding-left:200px">"Welcome {{ username }}!"</h1>
+    </div>
 </div>
 {% endblock %}
 
 {% block scripts %}
+    <script type="text/javascript">
+    $(document).ready(function() {
+     
+        // Navigate to author or method page on select
+        $("#view_container").click(function(){
+            console.log("View Container selected.")
+            var selected = $("#container").val();
+            document.location = '/container/' + selected
+        });
+
+    });
+    </script>
 {% endblock %}
