+++ tunel/templates/messages.html
@@ -1,0 +1,14 @@
+<div class='row'>
+    <div class='col-md-3'>
+    {% with messages = get_flashed_messages() %}
+        {% if messages %}
+        <hr>
+        <p id="messages" class="alert alert-info">
+        {% for message in messages %}
+            {{ message | safe }} </br>
+        {% endfor %}
+        </p>
+        {% endif %}
+    {% endwith %}
+    </div>
+</div>
