+++ tunel/templates/terminal.html
@@ -78,9 +78,6 @@
 
 <div class="terminal-window">
   <header>
-    <div class="button green"></div>
-    <div class="button yellow"></div>
-    <div class="button red"></div>
   </header>
   <section class="terminal">
     <div class="history"></div>
