+++ tunel/utils.py
@@ -1,6 +1,32 @@
-''' utils.py
+'''
 
-Utility functions for singularity-nginx
+Copyright (c) 2018, Vanessa Sochat
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+* Redistributions of source code must retain the above copyright notice, this
+  list of conditions and the following disclaimer.
+
+* Redistributions in binary form must reproduce the above copyright notice,
+  this list of conditions and the following disclaimer in the documentation
+  and/or other materials provided with the distribution.
+
+* Neither the name of the copyright holder nor the names of its
+  contributors may be used to endorse or promote products derived from
+  this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 '''
 
@@ -8,19 +34,17 @@ import os
 import re
 import subprocess
 import time
-from logman import bot
+from .logman import bot
 from glob import glob
-from singularity.cli import Singularity
-from singularity.utils import run_command
+from spython.main import Client
+from spython.utils import run_command
 
 
 def get_user():
     '''get_user will return the username. With a moder advanced app,
     we can hook this into a local database.
     '''
-    username = run_command('whoami')
-    if isinstance(username,bytes):
-        username = username.decode('utf-8')
+    username = run_command('whoami')['message'][0]
     return username.strip('\n').capitalize()
 
 
@@ -89,7 +113,7 @@ def get_container_labels(image_path,cli=None):
     :param cli: a client. Instantiated if not provided
     '''
     if cli is None:
-        cli = Singularity()
+        cli = Client
     return cli.get_labels(image_path)
 
 
@@ -97,7 +121,7 @@ def run_container(image_path,args=None,cli=None):
     '''run the container with one or more args
     '''
     if cli is None:
-        cli = Singularity()
+        cli = Client
 
     return cli.run(image_path,args=args,contain=True)
