+++ tunel/views/__init__.py
@@ -1,0 +1,3 @@
+from .api import *
+from .auth import *
+from .main import *
